backend/requirements.txt
fastapi
uvicorn[standard]
openai
langchain[openai,docarray]
langgraph
pandas
python-dotenv
tiktoken
chromadb
faiss-cpu